update request demo page
