--- src/components/Layout/index.js
@@ -1,3 +1,0 @@
-import Layout from './Layout'
-
-export default Layout
