+++ src/pages/index.js
@@ -124,9 +124,9 @@ function IndexPage() {
                                                 className="modalClose"
                                             />
                                         </Modal>
-                                        <a href="/request_demo">
-                                            <Button type="primary" size="large" className="requestDemo">
-                                                Request Demo
+                                        <a href="/schedule-demo">
+                                            <Button type="primary" size="large" className="scheduleDemo">
+                                                Schedule Demo
                                             </Button>
                                         </a>
                                     </div>
